+++ forui/test/src/widgets/bottom_navigation_bar_golden_test.dart
@@ -1,6 +1,7 @@
 @Tags(['golden'])
 library;
 
+import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 import 'package:forui/forui.dart';
@@ -9,7 +10,7 @@ import '../test_scaffold.dart';
 void main() {
   group('FBottomNavigationBar', () {
     for (final (name, theme, background) in TestScaffold.themes) {
-      testWidgets('$name with FBottomNavigationBar', (tester) async {
+      testWidgets('forui icon - $name', (tester) async {
         await tester.pumpWidget(
           TestScaffold(
             data: theme,
@@ -44,7 +45,52 @@ void main() {
 
         await expectLater(
           find.byType(TestScaffold),
-          matchesGoldenFile('bottom-navigation-bar/$name-bottom-navigation-bar.png'),
+          matchesGoldenFile('bottom-navigation-bar/forui-icon-$name.png'),
+        );
+      });
+
+      testWidgets('custom icon - $name', (tester) async {
+        Widget icon(BuildContext _, FBottomNavigationBarData data, Widget? __) => Container(
+          height: data.itemStyle.iconSize,
+          width: data.itemStyle.iconSize,
+          color: data.selected ? data.itemStyle.activeIconColor : data.itemStyle.inactiveIconColor,
+        );
+
+        await tester.pumpWidget(
+          TestScaffold(
+            data: theme,
+            background: background,
+            child: FBottomNavigationBar(
+              index: 2,
+              children: [
+                FBottomNavigationBarItem.customIcon(
+                  iconBuilder: icon,
+                  label: 'Home',
+                ),
+                FBottomNavigationBarItem.customIcon(
+                  iconBuilder: icon,
+                  label: 'Browse',
+                ),
+                FBottomNavigationBarItem.customIcon(
+                  iconBuilder: icon,
+                  label: 'Radio',
+                ),
+                FBottomNavigationBarItem.customIcon(
+                  iconBuilder: icon,
+                  label: 'Library',
+                ),
+                FBottomNavigationBarItem.customIcon(
+                  iconBuilder: icon,
+                  label: 'Search',
+                ),
+              ],
+            ),
+          ),
+        );
+
+        await expectLater(
+          find.byType(TestScaffold),
+          matchesGoldenFile('bottom-navigation-bar/custom-icon-$name.png'),
         );
       });
     }