+++ samples/lib/main.dart
@@ -42,6 +42,7 @@ class _AppRouter extends RootStackRouter {
         AutoRoute(path: '/avatar/invalid', page: AvatarInvalidRoute.page),
         AutoRoute(path: '/badge/default', page: BadgeRoute.page),
         AutoRoute(path: '/bottom-navigation-bar/default', page: BottomNavigationBarRoute.page),
+        AutoRoute(path: '/bottom-navigation-bar/custom', page: CustomBottomNavigationBarRoute.page),
         AutoRoute(path: '/button/text', page: ButtonTextRoute.page),
         AutoRoute(path: '/button/icon', page: ButtonIconRoute.page),
         AutoRoute(path: '/button/only-icon', page: ButtonOnlyIconRoute.page),